requirements.txt
google-api-python-client
google-auth
google-auth-httplib2
google-auth-oauthlib
gspread_asyncio
kivy
